- refactor views, components
