+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import './index.css';
-import App from './components/App/App';
+import App from './App';
 import reportWebVitals from './reportWebVitals';
 import 'react-dates/initialize';
 import 'react-dates/lib/css/_datepicker.css';
