+++ src/router.js
@@ -2,8 +2,8 @@
 import React from 'react';
 import {BrowserRouter as Router, Switch, Route} from 'react-router-dom';
 
-import Home from './components/Home';
-import NewOrder from './components/NewOrder';
+import Home from './views/Home';
+import NewOrder from './views/NewOrder';
 
 const router = () => {
   return (
